* upstream/master: (1175 commits)
  support server side js and arbitrary assets proper chaining for emoji so other hooks can be called
  ensure directory at the correct spot
  when there are no providers do not blow up
  plugins should not be checked in auto generated files should not be checked in
  pretty text would blow up on an <a tag with no href
  working plugin interface for custom openid auth, custom css and custom js
  odd formatting
  Use allow_uncategorized_topics site setting in categoryChoose
  Add indexes to help IncomingLinksReport, but not much. Refactor to limit result set size, but it doesn't help perf much either. SELECT COUNT(DISTINCT ...) is still slow.
  Add rake task multisite:rollback
  FIX: Back button broke when clicking @mentions
  Front-end Optimizations
  minor activation copy improvement
  oops background-color not color
  Remove redundant 'are'.
  Fix assets:precompile failure from missing @include
  prettify staff bg slightly
  make topic title icons properly sized
  we don't need a whole extra group for 1 pref
  simplify @mention CSS, add $quote-background color
  ...

Conflicts:
	Gemfile
	Gemfile.lock

+++ app/controllers/admin/groups_controller.rb
@@ -1,40 +1,58 @@
 class Admin::GroupsController < Admin::AdminController
   def index
-    groups = Group.order(:name).all
+    groups = Group.order(:name).to_a
     render_serialized(groups, BasicGroupSerializer)
   end
 
   def refresh_automatic_groups
     Group.refresh_automatic_groups!
-    render json: "ok"
+    render json: success_json
   end
 
   def users
     group = Group.find(params[:group_id].to_i)
-    render_serialized(group.users.limit(100).to_a, BasicUserSerializer)
+    render_serialized(group.users.order('username_lower asc').limit(200).to_a, BasicUserSerializer)
   end
 
   def update
     group = Group.find(params[:id].to_i)
-    render_json_error if group.automatic
-    group.usernames = params[:group][:usernames]
-    group.name = params[:group][:name] if params[:name]
-    group.save!
-    render json: "ok"
+    if group.automatic
+      can_not_modify_automatic
+    else
+      group.usernames = params[:group][:usernames]
+      group.name = params[:group][:name] if params[:group][:name]
+      if group.save
+        render json: success_json
+      else
+        render_json_error group
+      end
+    end
   end
 
   def create
     group = Group.new
-    group.name = params[:group][:name]
+    group.name = params[:group][:name].strip
     group.usernames = params[:group][:usernames] if params[:group][:usernames]
-    group.save!
-    render_serialized(group, BasicGroupSerializer)
+    if group.save
+      render_serialized(group, BasicGroupSerializer)
+    else
+      render_json_error group
+    end
   end
 
   def destroy
     group = Group.find(params[:id].to_i)
-    render_json_error if group.automatic
-    group.destroy
-    render json: "ok"
+    if group.automatic
+      can_not_modify_automatic
+    else
+      group.destroy
+      render json: success_json
+    end
+  end
+
+  protected
+
+  def can_not_modify_automatic
+    render json: {errors: I18n.t('groups.errors.can_not_modify_automatic')}, status: 422
   end
 end
